refactor: extract scheduling into AppstrateModule
Move schedule CRUD, BullMQ worker, and routes into
apps/api/src/modules/scheduling/. Remove hardcoded init/shutdown
from boot.ts and shutdown.ts. Register as built-in module.
