+++ apps/api/src/lib/shutdown.ts
@@ -5,7 +5,6 @@ import { logger } from "./logger.ts";
 import { shutdownInfra } from "../infra/index.ts";
 import { shutdownModules } from "./modules/index.ts";
 import { hasRedis } from "../infra/mode.ts";
-import { shutdownScheduleWorker } from "../services/scheduler.ts";
 import {
   getInFlightCount,
   waitForInFlight,
@@ -23,8 +22,7 @@ export function createShutdownHandler(setShuttingDown: () => void): () => Promis
     called = true;
     setShuttingDown();
 
-    logger.info("Shutdown initiated, stopping scheduler, webhook worker, and sidecar pool...");
-    await shutdownScheduleWorker();
+    logger.info("Shutdown initiated, stopping sidecar pool...");
     await getOrchestrator().shutdown();
 
     // Unsubscribe from cancel channel before draining to avoid processing
